pyproject.toml
[project]
name = "livex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.5.3",
    "channels-redis>=4.3.0",
    "django>=5.2.7",
    "django-celery-beat>=2.8.1",
    "django-redis>=6.0.0",
    "python-decouple>=3.8",
    "uvicorn[standard]>=0.38.0",
    "whitenoise>=6.11.0",
]
